+++ packages/core/src/utilities/fnv1aHash.ts
@@ -1,0 +1,14 @@
+/**
+ * Generates a hash for a string using FNV-1a algorithm
+ * @param str - string to hash
+ * @returns the hashed string in base 36
+ */
+export default function fnv1aHash(str: string): string {
+  let hash = 0x811c9dc5;
+  for (let i = 0; i < str.length; i++) {
+    hash ^= str.charCodeAt(i);
+    hash +=
+      (hash << 1) + (hash << 4) + (hash << 7) + (hash << 8) + (hash << 24);
+  }
+  return (hash >>> 0).toString(36);
+}