+++ packages/tools/src/tools/displayTools/Surface/surfaceDisplay.ts
@@ -11,10 +11,9 @@ import removeSurfaceFromElement from './removeSurfaceFromElement';
 import addOrUpdateSurfaceToElement from './addOrUpdateSurfaceToElement';
 import { getSegmentation } from '../../../stateManagement/segmentation/getSegmentation';
 import { getColorLUT } from '../../../stateManagement/segmentation/getColorLUT';
-import { canComputeRequestedRepresentation } from '../../../stateManagement/segmentation/polySeg/canComputeRequestedRepresentation';
-import { computeAndAddSurfaceRepresentation } from '../../../stateManagement/segmentation/polySeg/Surface/computeAndAddSurfaceRepresentation';
+import { getPolySeg } from '../../../config';
+import { computeAndAddRepresentation } from '../../../utilities/segmentation/computeAndAddRepresentation';
 
-const { ViewportType } = Enums;
 /**
  * It removes a segmentation representation from the tool group's viewports and
  * from the segmentation state
@@ -68,19 +67,38 @@ async function render(
 
   if (
     !SurfaceData &&
-    canComputeRequestedRepresentation(segmentationId, Representations.Surface)
+    getPolySeg()?.canComputeRequestedRepresentation(
+      segmentationId,
+      Representations.Surface
+    )
   ) {
     // we need to check if we can request polySEG to convert the other
     // underlying representations to Surface
-    SurfaceData = await computeAndAddSurfaceRepresentation(segmentationId, {
-      viewport,
-    });
+    const polySeg = getPolySeg();
+
+    SurfaceData = await computeAndAddRepresentation(
+      segmentationId,
+      Representations.Surface,
+      () => polySeg.computeSurfaceData(segmentationId, { viewport }),
+      () => polySeg.updateSurfaceData(segmentationId, { viewport })
+    );
 
     if (!SurfaceData) {
       throw new Error(
-        `No Surface data found for segmentationId ${segmentationId}.`
+        `No Surface data found for segmentationId ${segmentationId} even we tried to compute it`
       );
     }
+  } else if (!SurfaceData && !getPolySeg()) {
+    console.debug(
+      `No surface data found for segmentationId ${segmentationId} and PolySeg add-on is not configured. Unable to convert from other representations to surface. Please register PolySeg using cornerstoneTools.init({ addons: { polySeg } }) to enable automatic conversion.`
+    );
+  }
+
+  if (!SurfaceData) {
+    console.warn(
+      `No Surface data found for segmentationId ${segmentationId}. Skipping render.`
+    );
+    return;
   }
 
   const { geometryIds } = SurfaceData;