+++ packages/tools/src/tools/segmentation/strategies/compositions/preview.ts
@@ -28,9 +28,9 @@ export default {
   [StrategyCallbacks.Preview]: function (
     operationData: InitializedOperationData
   ) {
-    const { previewColors, strategySpecificConfiguration, enabledElement } =
-      operationData;
-    if (!previewColors || !strategySpecificConfiguration) {
+    const { previewColors, configuration, enabledElement } = operationData;
+
+    if (!previewColors || !configuration) {
       return;
     }
 
@@ -38,16 +38,19 @@ export default {
     if (operationData.preview) {
       delete operationData.preview;
     }
-    delete strategySpecificConfiguration.centerSegmentIndex;
+
+    delete configuration.centerSegmentIndex;
 
     // Now generate a normal preview as though the user had clicked, filled, released
     this.onInteractionStart?.(enabledElement, operationData);
+
     const preview = this.fill(enabledElement, operationData);
     if (preview) {
       preview.isPreviewFromHover = true;
       operationData.preview = preview;
       this.onInteractionEnd?.(enabledElement, operationData);
     }
+
     return preview;
   },
 
@@ -125,6 +128,7 @@ export default {
     if (previewSegmentIndex === undefined) {
       return;
     }
+
     const segmentIndex = preview?.segmentIndex ?? operationData.segmentIndex;
     if (!previewVoxelManager || previewVoxelManager.modifiedSlices.size === 0) {
       return;
@@ -175,7 +179,7 @@ export default {
     previewVoxelManager.forEach(callback);
 
     // Primarily rejects back to zero, so use 0 as the segment index - even
-    // if somtimes it modifies the data to other values on reject.
+    // if sometimes it modifies the data to other values on reject.
     triggerSegmentationDataModified(
       operationData.segmentationId,
       previewVoxelManager.getArrayOfModifiedSlices(),